% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth_tidal_filter.R
\name{butterworth_tidal_filter}
\alias{butterworth_tidal_filter}
\title{Return non-tidal signal in data after Rulh and Simpson (2005)}
\usage{
butterworth_tidal_filter(t, x)
}
\arguments{
\item{t}{time for x (time, POSIXct)}

\item{x}{any quantity, for example discharge (double)}
}
\value{
non-tidal signal in x with data affected by filter ringing removed
}
\description{
Applies a Butterworth filter with a 30-hour stop period and a 40-hour pass period
}
\examples{
t <- realTimeloads::ExampleData$Height$time
x <- realTimeloads::ExampleData$Height$Height_m
xf <- butterworth_tidal_filter(t,x)
}
\references{
Ruhl and Simpson (2005) https://pubs.usgs.gov/sir/2005/5004/sir20055004.pdf
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
