% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd2sal.R
\name{ctd2sal}
\alias{ctd2sal}
\title{Compute salinity (PSU) from conductivity, water temperature, and depth}
\usage{
ctd2sal(cond, temp, dbar)
}
\arguments{
\item{cond}{Conductance (microS/cm)}

\item{temp}{Water temperature (degrees C)}

\item{dbar}{Pressure (dBar) or water depth (m)}
}
\value{
Salinity in PSU
}
\description{
Computes salinity from conductivity, water temperature, and depth.
}
\section{Warning}{

If specific conductivity is returned from the sonde, the temperature at which specific conductivity is computed should be utilized
}

\examples{
Sonde <- realTimeloads::ExampleData$Sonde
S <- ctd2sal(Sonde$Conductivity_uS_per_cm,Sonde$Water_Temperature_degC,Sonde$Pressure_dbar)
}
\references{
UNESCO technical paper 44, 1983, with revisions to fresh water from Chen and Millero, 1986.

Author modified matlab code from David Schoellhamer
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
