% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleCode.R
\name{ExampleCode}
\alias{ExampleCode}
\title{Computes sediment load per guideline from ExampleData}
\usage{
ExampleCode()
}
\value{
list with data frames of estimated concentration and flux along with data used in regression and surrogate timeseries
}
\description{
Computes sediment load per guideline from ExampleData
}
\examples{
\donttest{
Output <- ExampleCode()
}
}
\references{
Livsey, D.N. (in review). National Industry Guidelines for hydrometric monitoring–Part 12: Application of acoustic Doppler velocity meters to measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia.
}
\seealso{
\code{\link{realTimeloads}} Package help file
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
