% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleCodeSCI.R
\name{ExampleCodeSCI}
\alias{ExampleCodeSCI}
\title{Computes sediment load from optical and acoustic backscatter measurements}
\usage{
ExampleCodeSCI()
}
\value{
total load with uncertainty computed from estimates of concentration from SCI
}
\description{
Computes sediment load per guideline from optical and acoustic backscatter measurements combined to the "Sediment Composition Index" (SCI) per Livsey et al (2023)
}
\examples{
\donttest{
Output <- ExampleCodeSCI()
}
}
\references{
Livsey, D.N. (in review). National Industry Guidelines for hydrometric monitoring–Part 12: Application of acoustic Doppler velocity meters to measure suspended-sediment load. Bureau of Meteorology. Melbourne, Australia.
}
\seealso{
\code{\link{realTimeloads}} Package help file
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
