% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data_with_arima.R
\name{impute_data_with_arima}
\alias{impute_data_with_arima}
\title{Returns x with gaps imputed using forecast::auto.arima with optional uncertainty estimation using Monte Carlo resampling}
\usage{
impute_data_with_arima(time, x, Xreg = NULL, ti = NULL, MC = 1, ptrain = 0.8)
}
\arguments{
\item{time}{time for x (time, POSIXct)}

\item{x}{any quantity (double)}

\item{Xreg}{additional predictors for ARIMA}

\item{ti}{time vector for interpolation (time, POSIXct)}

\item{MC}{number of Monte Carlo simulations for uncertainty estimation}

\item{ptrain}{proportion of data used for training and testing model}
}
\value{
list with x imputed at time or ti, if given. Uncertainty estimated from Monte Carlo simulations
}
\description{
Returns x with gaps imputed using forecast::auto.arima with optional uncertainty estimation using Monte Carlo resampling. Uncertainty on imputed data is estimated using using Monte Carlo (MC) resampling adapting methods of Rustomji and Wilkinson (2008). forecast::auto.arima searches for best ARIMA model and allows for step-changes in the ARIMA model of x(Xreg)
}
\note{
For infilling missing suspended-sediment concentration data (e.g., TSS or SSC) I suggest using the Froude number and discharge for Xreg. For tidally affected sites, consider using tidally filtered discharge.
}
\examples{
# Impute non-tidal data
time <- realTimeloads::ExampleData$Sediment_Samples$time
xo <- realTimeloads::ExampleData$Sediment_Samples$SSCxs_mg_per_liter
Q <- realTimeloads::ExampleData$Discharge$Discharge_m_cubed_per_s
idata <- sample(1:length(xo),round(length(xo)*0.5),replace=FALSE)
x <- rep(NA,length(xo))
x[idata] <- xo[idata] # simulated samples
Output <- impute_data_with_arima(time,x,Xreg = Q)

}
\references{
Rustomji, P., & Wilkinson, S. N. (2008). Applying bootstrap resampling to quantify uncertainty in fluvial suspended sediment loads estimated using rating curves. Water resources research, 44(9).

Hyndman R, Athanasopoulos G, Bergmeir C, Caceres G, Chhay L, O'Hara-Wild M, Petropoulos F, Razbash S, Wang E, Yasmeen F (2023). forecast: Forecasting functions for time series and linear models. R package version 8.21.1, https://pkg.robjhyndman.com/forecast/.

Hyndman RJ, Khandakar Y (2008). “Automatic time series forecasting: the forecast package for R.” Journal of Statistical Software, 26(3), 1–22. doi:10.18637/jss.v027.i03.
}
\author{
Daniel Livsey (2023) ORCID: 0000-0002-2028-6128
}
