% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interpolation_with_time_limit.R
\name{linear_interpolation_with_time_limit}
\alias{linear_interpolation_with_time_limit}
\title{Linearly interpolate timeseries t(x) onto new timesetep ti}
\usage{
linear_interpolation_with_time_limit(t, x, ti, threshold)
}
\arguments{
\item{t}{time for x (time, POSIXct)}

\item{x}{any quantity, for example discharge (double)}

\item{ti}{time where t(x) will be interpolated to (time, POSIXct)}

\item{threshold}{maximum duration where interpolation is allowed (hours)}
}
\value{
a data frame with time (ti), x interpolated from t(x) onto ti, and logical (ibad) if interpolation exceeded threshold
}
\description{
Linear interpolation limited by time since previous or following reading
}
\examples{
InputData <- realTimeloads::ExampleData
ADCP <- InputData$ADCP
Height <- InputData$Height
# Interpolate river height to ADCP time
t <- realTimeloads::ExampleData$Height$time
x <- realTimeloads::ExampleData$Height$Height_m
ti <-realTimeloads::ExampleData$ADCP$time
threshold <- 1
Output<- linear_interpolation_with_time_limit(t,x,ti,threshold)
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
