% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuation_of_sound_by_water.R
\name{attenuation_of_sound_by_water}
\alias{attenuation_of_sound_by_water}
\title{Compute attenuation of sound in water given frequency, temperature, and salinity}
\usage{
attenuation_of_sound_by_water(f, T, S)
}
\arguments{
\item{f}{frequency of sound (Hz)}

\item{T}{Water temperature (degrees C)}

\item{S}{Salinity (PSU)}
}
\value{
attenuation of sound in water (dB/m), divide by 20*log10(exp(1)) to convert to Nepers/m
}
\description{
Computes attenuation of sound in water per Ainslie and McColm (1998)
}
\examples{
InputData <- realTimeloads::ExampleData
f <- InputData$ADCP$Accoustic_Frequency_kHz*1000
cond <-InputData$Sonde$Conductivity_uS_per_cm
temp <- InputData$Sonde$Water_Temperature_degC
dbar <- InputData$Sonde$Pressure_dbar
S <- ctd2sal(cond,temp,dbar)
aw <- attenuation_of_sound_by_water(f,temp,S) # dB/m
awNp <- attenuation_of_sound_by_water(f,temp,S)/(20*log10(exp(1))) # Np/m
}
\references{
Ainslie and McColm (1998): https://doi.org/10.1121/1.421258

Author modified matlab code from David Schoellhamer
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
