% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleCode.R
\name{ExampleCode}
\alias{ExampleCode}
\title{Computes sediment load per guideline from ExampleData}
\usage{
ExampleCode()
}
\value{
list with data frames of estimated concentration and flux along with data used in regression and surrogate timeseries
}
\description{
Computes sediment load per guideline from ExampleData
}
\examples{
\dontrun{
Output <- ExampleCode()
}
}
\references{
GUIDLINE
}
\seealso{
\code{\link{realTimeloads}} Package help file
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
