% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleCodeSCI.R
\name{ExampleCodeSCI}
\alias{ExampleCodeSCI}
\title{Computes sediment load from optical and acoustic backscatter measurements}
\usage{
ExampleCodeSCI()
}
\value{
total load with uncertainty computed from estimates of concentration from SCI
}
\description{
Computes sediment load per guideline from optical and acoustic backscatter measurements combined to the "Sediment Composition Index" (SCI) see Livsey et al (2023)
}
\examples{
\dontrun{
Output <- ExampleCodeSCI()
}
}
\references{
GUIDLINE
Livsey et al (2023): https://doi.org/10.1029/2022WR033982
}
\seealso{
\code{\link{realTimeloads}} Package help file
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
