% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_of_sound.R
\name{speed_of_sound}
\alias{speed_of_sound}
\title{Compute speed of sound in water given salinity, temperature, and depth}
\usage{
speed_of_sound(S, T, D)
}
\arguments{
\item{S}{Salinity (PSU)}

\item{T}{Water temperature (degrees C)}

\item{D}{Water depth (m) or pressure (dBar)}
}
\value{
Speed of sound in water (m/s)
}
\description{
Computes speed of sound in water per Del grosso (1974)
}
\examples{
InputData <- realTimeloads::ExampleData
Sonde<- InputData$Sonde
S <- ctd2sal(Sonde$Conductivity_uS_per_cm,Sonde$Water_Temperature_degC,Sonde$Pressure_dbar)
c <- speed_of_sound(S,Sonde$Water_Temperature_degC,Sonde$Pressure_dbar)
}
\references{
Del grosso (1974): https://doi.org/10.1121/1.1903388

Author modified matlab code from David Schoellhamer
}
\author{
Daniel Livsey, September (2023), livsey.daniel@gmail.com, ORCID: 0000-0002-2028-6128
}
